requirements.txt
pywinauto>=0.6.8
customtkinter>=5.0.0
pystray>=0.19.5
Pillow>=10.0.0
schedule>=1.1.0
psutil>=5.9.0
pyautogui>=0.9.53